The debian postrm cleanup should happen on purge
The socorro files (configs specifically) and user should only be removed on
purge.
https://wiki.debian.org/AccountHandlingInMaintainerScripts has some
good guidelines regarding account deletion.

+++ scripts/package/deb-after-remove.sh
@@ -1,6 +1,13 @@
 #! /bin/bash
+set -e
 INSTALL_PREFIX=/data
 
-if [ -d $INSTALL_PREFIX/socorro ]; then
-	rm -r $INSTALL_PREFIX/socorro
+if [ "$1" = "purge" ] ; then
+    if [ -d $INSTALL_PREFIX/socorro ]; then
+        rm -r $INSTALL_PREFIX/socorro
+    fi
+
+    if type deluser >/dev/null 2>&1; then
+        userdel -r socorro 2>&1 > /dev/null
+    fi
 fi
